requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
future==0.16.0
gast==0.2.0
grpcio==1.11.0
gym==0.10.5
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.1.5
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.1
mistune==0.8.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.14.3
olefile==0.45.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
pyglet==1.3.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
pywin32==223
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
tb-nightly==1.8.0a20180424
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
testpath==0.3.1
tf-nightly-gpu==1.8.0.dev20180329
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.0.8
wincertstore==0.2
